Fixed dead links in gallery and changed formatting
There is an offset issue in the first row of the gallery. Its a known
issue with the old version of bootstrap.css that we are using. I tried
to solve the issue by upgrading but this broke a few other things on the
site. Someone with a better knowledge of css should take a look a this.